Blueprints
Added the use of blueprints to the architecture.
